+++ src/js/mixin/internal/animate-slide.js
@@ -1,23 +1,19 @@
-import {addClass, append, assign, children, css, fastdom, height, includes, index, isVisible, noop, offset, parent, position, Promise, removeClass, Transition} from 'uikit-util';
-
-const targetClass = 'uk-animation-target';
+import {assign, children, css, fastdom, includes, index, isVisible, noop, offset, parent, position, Promise, Transition} from 'uikit-util';
 
 export default function (action, target, duration) {
 
     return new Promise(resolve =>
         requestAnimationFrame(() => {
-            addStyle();
 
             let nodes = children(target);
 
             // Get current state
             const currentProps = nodes.map(el => getProps(el, true));
-            const oldHeight = height(target);
+            const targetProps = css(target, ['height', 'padding']);
 
             // Cancel previous animations
             Transition.cancel(target);
             nodes.forEach(Transition.cancel);
-            removeClass(target, targetClass);
             reset(target);
 
             // Adding, sorting, removing nodes
@@ -33,20 +29,19 @@ export default function (action, target, duration) {
                 fastdom.flush();
 
                 // Get new state
-                const newHeight = height(target);
+                const targetPropsTo = css(target, ['height', 'padding']);
                 const [propsTo, propsFrom] = getTransitionProps(target, nodes, currentProps);
 
                 // Reset to previous state
-                addClass(target, targetClass);
                 nodes.forEach((el, i) => propsFrom[i] && css(el, propsFrom[i]));
-                css(target, {height: oldHeight, display: 'block'});
+                css(target, assign({display: 'block'}, targetProps));
 
                 // Start transitions on next frame
                 requestAnimationFrame(() => {
 
                     const transitions = nodes.map((el, i) =>
                             parent(el) === target && Transition.start(el, propsTo[i], duration, 'ease')
-                        ).concat(Transition.start(target, {height: newHeight}, duration, 'ease'));
+                        ).concat(Transition.start(target, targetPropsTo, duration, 'ease'));
 
                     Promise.all(transitions).then(() => {
                         nodes.forEach((el, i) => parent(el) === target && css(el, 'display', propsTo[i].opacity === 0 ? 'none' : ''));
@@ -118,30 +113,19 @@ function reset(el) {
         pointerEvents: '',
         position: '',
         top: '',
+        marginTop: '',
+        marginLeft: '',
+        transform: '',
         width: '',
         zIndex: ''
     });
-    removeClass(el, targetClass);
-    css(el, {height: '', display: ''});
+    css(el, {height: '', display: '', padding: ''});
 }
 
 function getPositionWithMargin(el) {
     const {height, width} = offset(el);
     const {top, left} = position(el);
+    const {marginLeft, marginTop} = css(el, ['marginTop', 'marginLeft']);
 
-    return {top, left, height, width};
-}
-
-let style;
-
-function addStyle() {
-    if (style) {
-        return;
-    }
-    style = !!append(document.head, `<style>
-        .${targetClass} > * {
-            margin-top: 0 !important;
-            transform: none !important;
-        }
-    </style>`);
+    return {top, left, height, width, marginLeft, marginTop, transform: ''};
 }