+++ src/js/mixin/modal.js
@@ -185,6 +185,10 @@ export default {
 
         show() {
 
+            if (this.isToggled()) {
+                return Promise.resolve();
+            }
+
             if (this.container && parent(this.$el) !== this.container) {
                 append(this.container, this.$el);
                 return new Promise(resolve =>
@@ -198,6 +202,11 @@ export default {
         },
 
         hide() {
+
+            if (!this.isToggled()) {
+                return Promise.resolve();
+            }
+
             return this.toggleElement(this.$el, false, animate(this));
         }